+++ __mocks__/axios.ts
@@ -1,7 +1,7 @@
 export const mockedAxios = {
   create: () => mockedAxios,
   nextResponse: {},
-  request: () => Promise.resolve(mockedAxios.nextResponse),
+  request: () => Promise.resolve(mockedAxios.nextResponse)
 };
 
 export default mockedAxios;